requirements.txt
python-binance==1.0.28
pandas==1.5.3
numpy==1.24.3
python-telegram-bot==13.15
ta==0.10.2
plotly==5.14.1
kaleido==0.2.1
matplotlib==3.7.1
python-dotenv==1.0.0
scikit-learn==1.3.0
